Feat: config.py 교체
- 기존 config.py와 config2.py 파일 사용사례 교체
